Tests for Avaruus: explictly cast `size()` to `int`.

+++ code/avaruus/cpp/tests/test_get_first_token.cpp
@@ -30,7 +30,7 @@ TEST(get_first_token_must_work_properly, empty)
     std::string_view empty { "" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(empty.data(), empty.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(empty.data(), static_cast<int>(empty.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -42,7 +42,7 @@ TEST(get_first_token_must_work_properly, space)
     std::string_view space { " " };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space.data(), space.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space.data(), static_cast<int>(space.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -54,7 +54,7 @@ TEST(get_first_token_must_work_properly, a)
     std::string_view a { "a" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a.data(), a.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a.data(), static_cast<int>(a.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -66,7 +66,7 @@ TEST(get_first_token_must_work_properly, hash)
     std::string_view hash { "#" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(hash.data(), hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(hash.data(), static_cast<int>(hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -78,7 +78,7 @@ TEST(get_first_token_must_work_properly, newline)
     std::string_view newline { "\n" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(newline.data(), newline.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(newline.data(), static_cast<int>(newline.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -90,7 +90,7 @@ TEST(get_first_token_must_work_properly, space_a)
     std::string_view space_a { " a" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a.data(), space_a.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a.data(), static_cast<int>(space_a.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 3); // Searching for next token can begin from Fortran index 3.
@@ -102,7 +102,7 @@ TEST(get_first_token_must_work_properly, space_hash)
     std::string_view space_hash { " #" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_hash.data(), space_hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_hash.data(), static_cast<int>(space_hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -114,7 +114,7 @@ TEST(get_first_token_must_work_properly, a_space)
     std::string_view a_space { "a " };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space.data(), a_space.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space.data(), static_cast<int>(a_space.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -126,7 +126,7 @@ TEST(get_first_token_must_work_properly, a_hash)
     std::string_view a_hash { "a#" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_hash.data(), a_hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_hash.data(), static_cast<int>(a_hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -138,7 +138,7 @@ TEST(get_first_token_must_work_properly, a_newline)
     std::string_view a_newline { "a\n" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_newline.data(), a_newline.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_newline.data(), static_cast<int>(a_newline.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -150,7 +150,7 @@ TEST(get_first_token_must_work_properly, newline_a)
     std::string_view newline_a { "\na" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(newline_a.data(), newline_a.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(newline_a.data(), static_cast<int>(newline_a.size()), &next_i, &token_sz);
     std::string_view expected_token { "" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, -1); // No token found!
@@ -162,7 +162,7 @@ TEST(get_first_token_must_work_properly, space_a_space)
     std::string_view space_a_space { " a " };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a_space.data(), space_a_space.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a_space.data(), static_cast<int>(space_a_space.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 3); // Searching for next token can begin from Fortran index 3.
@@ -174,7 +174,7 @@ TEST(get_first_token_must_work_properly, space_a_hash)
     std::string_view space_a_hash { " a#" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a_hash.data(), space_a_hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(space_a_hash.data(), static_cast<int>(space_a_hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 3); // Searching for next token can begin from Fortran index 3.
@@ -186,7 +186,7 @@ TEST(get_first_token_must_work_properly, a_space_a)
     std::string_view a_space_a { "a a" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space_a.data(), a_space_a.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space_a.data(), static_cast<int>(a_space_a.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -198,7 +198,7 @@ TEST(get_first_token_must_work_properly, a_space_hash)
     std::string_view a_space_hash { "a #" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space_hash.data(), a_space_hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(a_space_hash.data(), static_cast<int>(a_space_hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "a" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 2); // Searching for next token can begin from Fortran index 2.
@@ -210,7 +210,7 @@ TEST(get_first_token_must_work_properly, abc)
     std::string_view abc { "abc" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc.data(), abc.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc.data(), static_cast<int>(abc.size()), &next_i, &token_sz);
     std::string_view expected_token { "abc" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 4); // Searching for next token can begin from Fortran index 4.
@@ -222,7 +222,7 @@ TEST(get_first_token_must_work_properly, abc_space)
     std::string_view abc_space { "abc " };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_space.data(), abc_space.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_space.data(), static_cast<int>(abc_space.size()), &next_i, &token_sz);
     std::string_view expected_token { "abc" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 4); // Searching for next token can begin from Fortran index 4.
@@ -234,7 +234,7 @@ TEST(get_first_token_must_work_properly, abc_hash)
     std::string_view abc_hash { "abc#" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_hash.data(), abc_hash.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_hash.data(), static_cast<int>(abc_hash.size()), &next_i, &token_sz);
     std::string_view expected_token { "abc" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 4); // Searching for next token can begin from Fortran index 4.
@@ -246,7 +246,7 @@ TEST(get_first_token_must_work_properly, abc_newline)
     std::string_view abc_newline { "abc\n" };
     int next_i = 0;
     int token_sz = 0;
-    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_newline.data(), abc_newline.size(), &next_i, &token_sz);
+    char* token = FortranCInterface_MODULE_(string_mod, get_first_token, string_mod, get_first_token)(abc_newline.data(), static_cast<int>(abc_newline.size()), &next_i, &token_sz);
     std::string_view expected_token { "abc" };
     ASSERT_EQ(expected_token.compare(token), 0);
     ASSERT_EQ(next_i, 4); // Searching for next token can begin from Fortran index 4.
